feat: 게임 종료시 result 뷰로 가도록 구현
