+++ src/main/resources/templates/play.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <title>Title</title>
+    <title>게임 화면</title>
 </head>
 <body>
     {{turn}} 차례입니다.
