+++ src/main/resources/templates/result.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>게임 결과</title>
+</head>
+<body>
+    승자: {{winner}} <br/>
+    백팀 점수: {{whiteScore}} <br/>
+    흑팀 점수: {{blackScore}} <br/>
+</body>
+</html>
